# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Varese Life API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SOURCES
...
  /browseract/webhook/{sourceId}/{sourceUrlId}:
    post:
      operationId: browseractWebhook
      tags: [browseract]
      summary: BrowserAct webhook receiver
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: integer
        - name: sourceUrlId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /browseract/task/{taskId}/status:
    get:
      operationId: getBrowseractTaskStatus
      tags: [browseract]
      summary: Get BrowserAct task status (polling fallback)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatusResponse"

  # EVENTS
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: sourceId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /events/{id}/rewrite:
    post:
      operationId: rewriteEvent
      tags: [events]
      summary: Rewrite event description with AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rewritten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/rescrape:
    post:
      operationId: rescrapeEvent
      tags: [events]
      summary: Re-scrape event from source URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rescrape triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeResponse"
  /events/bulk:
    post:
      operationId: bulkUpdateEvents
      tags: [events]
      summary: Bulk update event statuses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateEventsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /events/extract-screenshot:
    post:
      operationId: extractFromScreenshot
      tags: [events]
      summary: Extract event from screenshot images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractScreenshotRequest"
      responses:
        "200":
          description: Extracted fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedEventFields"
  /events/extract-text:
    post:
      operationId: extractFromText
      tags: [events]
      summary: Extract event from free text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractTextRequest"
      responses:
        "200":
          description: Extracted fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedEventFields"
  /events/extract-url:
    post:
      operationId: extractFromUrl
      tags: [events]
      summary: Extract event from URL (BrowserAct or Claude fallback)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractUrlRequest"
      responses:
        "200":
          description: Extracted fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedEventFields"

  # PROMPTS
...
  /prompts/{id}:
    get:
      operationId: getPrompt
      tags: [prompts]
      summary: Get prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
    put:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromptRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}/reset:
    post:
      operationId: resetPrompt
      tags: [prompts]
      summary: Reset prompt to default
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"

  # NEWSLETTERS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        isSuperadmin:
          type: boolean
      required: [id, email, role, isSuperadmin]
...
    ScrapeResponse:
      type: object
      properties:
        message:
          type: string
        taskIds:
          type: array
          items:
            type: string
      required: [message]
...
    TaskStatusResponse:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
        result:
          type: object
          nullable: true
      required: [taskId, status]
    Event:
      type: object
      properties:
        id:
          type: integer
        sourceId:
          type: integer
          nullable: true
        title:
          type: string
        dateStart:
          type: string
          nullable: true
        dateEnd:
          type: string
          nullable: true
        dateDisplay:
          type: string
          nullable: true
        dateParseConfidence:
          type: string
        recurring:
          type: boolean
        location:
          type: string
          nullable: true
        descriptionRaw:
          type: string
          nullable: true
        descriptionClean:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
        status:
          type: string
        includedInNewsletter:
          type: boolean
        sourceType:
          type: string
        scraper:
          type: string
          nullable: true
        createdAt:
          type: string
        sourceName:
          type: string
          nullable: true
      required: [id, title, dateParseConfidence, recurring, status, includedInNewsletter, sourceType, createdAt]
    EventsListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [events, total, page, limit]
    CreateEventRequest:
      type: object
      properties:
        sourceId:
          type: integer
          nullable: true
        title:
          type: string
        dateStart:
          type: string
          nullable: true
        dateEnd:
          type: string
          nullable: true
        dateDisplay:
          type: string
          nullable: true
        dateParseConfidence:
          type: string
        recurring:
          type: boolean
        location:
          type: string
          nullable: true
        descriptionRaw:
          type: string
          nullable: true
        descriptionClean:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
        status:
          type: string
        sourceType:
          type: string
      required: [title]
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
        dateStart:
          type: string
          nullable: true
        dateEnd:
          type: string
          nullable: true
        dateDisplay:
          type: string
          nullable: true
        dateParseConfidence:
          type: string
        recurring:
          type: boolean
        location:
          type: string
          nullable: true
        descriptionRaw:
          type: string
          nullable: true
        descriptionClean:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
        status:
          type: string
        includedInNewsletter:
          type: boolean
    BulkUpdateEventsRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        action:
          type: string
      required: [ids, action]
    ExtractScreenshotRequest:
      type: object
      properties:
        images:
          type: array
          items:
            type: string
        sourceUrl:
          type: string
          nullable: true
      required: [images]
    ExtractTextRequest:
      type: object
      properties:
        text:
          type: string
        sourceUrl:
          type: string
          nullable: true
      required: [text]
    ExtractUrlRequest:
      type: object
      properties:
        url:
          type: string
      required: [url]
    ExtractedEventFields:
      type: object
      properties:
        title:
          type: string
          nullable: true
        dateStart:
          type: string
          nullable: true
        dateEnd:
          type: string
          nullable: true
        dateDisplay:
          type: string
          nullable: true
        dateParseConfidence:
          type: string
        recurring:
          type: boolean
        location:
          type: string
          nullable: true
        descriptionRaw:
          type: string
          nullable: true
        descriptionClean:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
      required: [dateParseConfidence, recurring]
    Prompt:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        label:
          type: string
        content:
          type: string
        description:
          type: string
        defaultContent:
          type: string
        updatedAt:
          type: string
      required: [id, key, label, content, description, defaultContent, updatedAt]
    UpdatePromptRequest:
      type: object
      properties:
        label:
          type: string
        content:
          type: string
        description:
          type: string
...