# lib/api-zod/src/generated/types/updatePromptRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Varese Life API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePromptRequest {
  label?: string;
  content?: string;
  description?: string;
}